requirements.txt
gradio
pandas
numpy
faiss-cpu
sentence-transformers
langchain
